ultra/requirements.txt
loguru==0.5.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
psycopg2==2.8.5
python-dateutil==2.8.1
ballcosmos==0.1.2
